requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.5
django-appconf==1.0.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-channels==0.7.0
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
envparse==0.2.0
future==0.18.2
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
loguru==0.5.3
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
oauthlib==3.1.1
phonenumbers==8.12.28
pilkit==2.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
reportlab==3.5.68
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
win32-setctime==1.0.3
xhtml2pdf==0.2.5
yarl==1.6.3
zope.interface==5.4.0
